requirements.txt
# AI Data Analyst - Dependencies
# Python 3.9+ recommended

# Web Framework
streamlit>=1.28.0

# OpenAI SDK
openai>=1.0.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.0.0
pandas>=2.0.0

# Numerical Computing (for embeddings & similarity)
numpy>=1.24.0
scikit-learn>=1.3.0

# Optional: Better PDF text extraction
# pdfplumber>=0.10.0
